requirements.txt
Django==2.2.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
django-allauth==0.39.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-environ==0.4.5